// apps/cms/src/payload-types.ts
export interface StockHistory {
  id: string;
  name: string;
  currency: string;
  amount: number;
  percent_change: number;
  volume: number;
  symbol: string;
  date: string;
  updatedAt: string;
  createdAt: string;
}